pyproject.toml
[tool.poetry]
name = "devreplay-pattern-generator"
version = "0.1.0"
description = ""
authors = ["tomoya-n <s266227@wakayama-u.ac.jp>"]
readme = "README.md"
packages = [
    { include = "devreplay-pattern-generator" }
]

[tool.poetry.dependencies]
python = "^3.12"
antlr4-python3-runtime = "4.8"
gitpython = "^3.1.43"
unidiff = "^0.7.5"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
